pyproject.toml
[tool.poetry]
name = "asyncapi-container"
version = "0.1.0"
description = ""
authors = ["Sergei Konik <s.konik.dev@gmail.com>"]
readme = "README.md"
packages = [{include = "asyncapi_container"}]

[tool.poetry.dependencies]
python = "^3.10"
attrs = "^23.2.0"
pydantic = "^1.10.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
